refactor adjust middle 3x3 row of hero text allignment
