+++ frontend/web/src/features/hostdashboard/website/WebsiteEditorPage.module.scss
@@ -700,6 +700,8 @@
     linear-gradient(180deg, rgba(248, 250, 252, 0.98), rgba(241, 245, 249, 0.98)),
     #f8fafc;
   display: grid;
+  align-items: center;
+  justify-items: center;
   padding: 4px;
 }
